requirements.txt
yfinance 
alpha_vantage 
pandas 
numpy 
logging
python-dotenv
lxml
matplotlib